Updated Content

+++ 14_SpringDataJPA/src/main/java/lk/ijse/spring/service/CustomerServiceImpl.java
@@ -27,7 +27,11 @@ public class CustomerServiceImpl  implements CustomerService {
     }
     @Override
     public void deleteCustomer(String id){
-        repo.deleteById(id);
+        if (repo.existsById(id)) {
+            repo.deleteById(id);
+        }else {
+            throw  new RuntimeException("Please check the customer ID ... No Such Customer ...!!");
+        }
     }
 
     @Override
